test(drawer): cover stacked drag in browser
